+++ apps/webapp/app/components/navigation/SideMenu.tsx
@@ -219,7 +219,7 @@ function SideMenu({
               </li>
               <li>
                 <a
-                  href="https://docs.trigger.dev/introduction#getting-started"
+                  href="https://docs.trigger.dev/getting-started"
                   rel="noreferrer"
                   target="_blank"
                   className={menuSmallLinkStyle}