+++ apps/webapp/app/services/workflows/registerWorkflow.server.ts
@@ -45,12 +45,23 @@ export class RegisterWorkflow {
     }
 
     if (validation.data.trigger.service !== "trigger") {
-      await this.#upsertExternalSource(
+      const source = await this.upsertSource(
         validation.data,
         organization,
         workflow,
         environment
       );
+
+      if (source && source.status === "READY") {
+        await this.#prismaClient.workflow.update({
+          where: {
+            id: workflow.id,
+          },
+          data: {
+            status: "READY",
+          },
+        });
+      }
     }
 
     await this.#upsertEventRule(
@@ -99,6 +110,18 @@ export class RegisterWorkflow {
     payload: WorkflowMetadata,
     organization: Organization
   ) {
+    const existingWorkflow = await this.#prismaClient.workflow.findUnique({
+      where: {
+        organizationId_slug: {
+          organizationId: organization.id,
+          slug,
+        },
+      },
+      select: {
+        id: true,
+      },
+    });
+
     const workflow = await this.#prismaClient.workflow.upsert({
       where: {
         organizationId_slug: {
@@ -130,10 +153,20 @@ export class RegisterWorkflow {
       },
     });
 
+    if (!existingWorkflow) {
+      await taskQueue.publish("SEND_INTERNAL_EVENT", {
+        id: workflow.id,
+        name: "workflow.created",
+        payload: {
+          id: workflow.id,
+        },
+      });
+    }
+
     return workflow;
   }
 
-  async #upsertExternalSource(
+  async upsertSource(
     payload: WorkflowMetadata,
     organization: Organization,
     workflow: Workflow,
@@ -197,6 +230,31 @@ export class RegisterWorkflow {
         });
 
         return schedulerSource;
+      }
+      case "SLACK_INTERACTION": {
+        if (!payload.trigger.source) {
+          return;
+        }
+
+        return this.#prismaClient.internalSource.upsert({
+          where: {
+            workflowId_environmentId: {
+              workflowId: workflow.id,
+              environmentId: environment.id,
+            },
+          },
+          update: {
+            source: payload.trigger.source,
+          },
+          create: {
+            organizationId: organization.id,
+            workflowId: workflow.id,
+            environmentId: environment.id,
+            source: payload.trigger.source,
+            status: "READY",
+            type: "SLACK",
+          },
+        });
       }
       default: {
         return;