+++ packages/trigger-providers/src/providers/slack/index.ts
@@ -15,6 +15,8 @@ export const slack = {
       "groups:write",
       "im:write",
       "mpim:write",
+      "chat:write.customize",
+      "reactions:write",
     ],
   },
   schemas,